Move abstract rule_base methods to optional interfaces
