+++ classes/local/rule/rule_base.php
@@ -17,12 +17,11 @@
 namespace tool_registrationrules\local\rule;
 
 use coding_exception;
-use MoodleQuickForm;
 use stdClass;
 use tool_registrationrules\local\rule_check_result;
 
 /**
- * Interface for registration rule subplugin classes.
+ * Base class for rule subplugin classes.
  *
  * @package    tool_registrationrules
  * @copyright 2024 Catalyst IT Europe {@link https://www.catalyst-eu.net}
@@ -34,7 +33,7 @@ use tool_registrationrules\local\rule_check_result;
  * @author    Lukas MuLu Müller <info@mulu.at>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-abstract class rule_base implements rule_interface {
+class rule_base {
 
     /** @var stdClass rule plugin instance config. */
     protected stdClass $instanceconfig;
@@ -213,28 +212,4 @@ abstract class rule_base implements rule_interface {
 
         return $this->result;
     }
-
-    /**
-     * Inject additional fields into the signup form for usage by the rule instance after submission.
-     *
-     * @param MoodleQuickForm $mform
-     * @return void
-     */
-    public function extend_form(MoodleQuickForm $mform): void {
-    }
-
-    /**
-     * Perform rule's checks applicable without any user input before the signup form is displayed.
-     *
-     * @return rule_check_result|null A rule_check_result object or null if check not applicable for this type.
-     */
-    abstract public function pre_data_check(): ?rule_check_result;
-
-    /**
-     * Perform rule's checks based on form input and user behaviour after signup form is submitted.
-     *
-     * @param array $data the data array from submitted form values.
-     * @return rule_check_result|null a rule_check_result object or null if check not applicable for this type.
-     */
-    abstract public function post_data_check(array $data): ?rule_check_result;
 }
