+++ rules/disposableemails/classes/rule.php
@@ -18,8 +18,8 @@ namespace registrationrule_disposableemails;
 
 use moodle_exception;
 use registrationrule_disposableemails\local\list_manager;
+use tool_registrationrules\local\rule\post_data_check;
 use tool_registrationrules\local\rule\rule_base;
-use tool_registrationrules\local\rule\rule_interface;
 use tool_registrationrules\local\rule_check_result;
 
 /**
@@ -34,24 +34,15 @@ use tool_registrationrules\local\rule_check_result;
  * @author    Lukas MuLu Müller <info@mulu.at>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class rule extends rule_base implements rule_interface {
-    /**
-     * Perform rule's checks applicable without any user input before the signup form is displayed.
-     *
-     * @return rule_check_result|null A rule_check_result object or null if check not applicable for this type.
-     */
-    public function pre_data_check(): ?rule_check_result {
-        return null;
-    }
-
+class rule extends rule_base implements post_data_check {
     /**
      * Perform rule's checks based on form input and user behaviour after signup form is submitted.
      *
      * @param array $data the data array from submitted form values.
-     * @return rule_check_result|null a rule_check_result object or null if check not applicable for this type.
+     * @return rule_check_result a rule_check_result object.
      * @throws moodle_exception
      */
-    public function post_data_check(array $data): ?rule_check_result {
+    public function post_data_check(array $data): rule_check_result {
         if (!array_key_exists('email', $data)) {
             return null;
         }
