+++ rules/hcaptcha/classes/rule.php
@@ -19,7 +19,9 @@ namespace registrationrule_hcaptcha;
 use coding_exception;
 use curl;
 use MoodleQuickForm;
+use tool_registrationrules\local\rule\extend_signup_form;
 use tool_registrationrules\local\rule\plugin_configurable;
+use tool_registrationrules\local\rule\post_data_check;
 use tool_registrationrules\local\rule_check_result;
 
 /**
@@ -35,7 +37,8 @@ use tool_registrationrules\local\rule_check_result;
  * @author    Lukas MuLu Müller <info@mulu.at>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class rule extends \tool_registrationrules\local\rule\rule_base implements plugin_configurable {
+class rule extends \tool_registrationrules\local\rule\rule_base
+implements extend_signup_form, plugin_configurable, post_data_check {
     /**
      * Inject additional fields into the signup form for usage by the rule instance after submission.
      *
@@ -65,10 +68,10 @@ class rule extends \tool_registrationrules\local\rule\rule_base implements plugi
      * Perform rule's checks based on form input and user behaviour after signup form is submitted.
      *
      * @param array $data the data array from submitted form values.
-     * @return rule_check_result|null a rule_check_result object or null if check not applicable for this type.
+     * @return rule_check_result a rule_check_result object.
      * @throws coding_exception
      */
-    public function post_data_check(array $data): ?rule_check_result {
+    public function post_data_check(array $data): rule_check_result {
         $hcaptchasecret = get_config('registrationrule_hcaptcha', 'hcaptcha_secret');
         $hcaptchasitekey = get_config('registrationrule_hcaptcha', 'hcaptcha_sitekey');
         // Return early if the config has not been set.
@@ -111,15 +114,6 @@ class rule extends \tool_registrationrules\local\rule\rule_base implements plugi
 
         // We got to this point so the captcha check passed.
         return $this->allow();
-    }
-
-    /**
-     * Perform rule's checks applicable without any user input before the signup form is displayed.
-     *
-     * @return rule_check_result|null A rule_check_result object or null if check not applicable for this type.
-     */
-    public function pre_data_check(): ?rule_check_result {
-        return null;
     }
 
     /**
