+++ rules/hibp/classes/rule.php
@@ -18,6 +18,7 @@ namespace registrationrule_hibp;
 
 use coding_exception;
 use curl;
+use tool_registrationrules\local\rule\post_data_check;
 use tool_registrationrules\local\rule_check_result;
 
 /**
@@ -31,15 +32,15 @@ use tool_registrationrules\local\rule_check_result;
  * @author    Lukas MuLu Müller <info@mulu.at>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class rule extends \tool_registrationrules\local\rule\rule_base {
+class rule extends \tool_registrationrules\local\rule\rule_base implements post_data_check {
     /**
      * Perform rule's checks based on form input and user behaviour after signup form is submitted.
      *
      * @param array $data the data array from submitted form values.
-     * @return rule_check_result|null a rule_check_result object or null if check not applicable for this type.
+     * @return rule_check_result a rule_check_result object.
      * @throws coding_exception
      */
-    public function post_data_check(array $data): ?rule_check_result {
+    public function post_data_check(array $data): rule_check_result {
         if (!isset($data['password'])) {
             return null;
         }
@@ -99,14 +100,5 @@ class rule extends \tool_registrationrules\local\rule\rule_base {
 
         // We got to this point so looks like we have not been pwned after all.
         return $this->allow();
-    }
-
-    /**
-     * Perform rule's checks applicable without any user input before the signup form is displayed.
-     *
-     * @return rule_check_result|null A rule_check_result object or null if check not applicable for this type.
-     */
-    public function pre_data_check(): ?rule_check_result {
-        return null;
     }
 }
